🔨 refactor: 중복 Entity, Repository 통합
