--- src/main/java/com/x1/groo/forest/common/domain/aggregate/AnnouncementEntity.java
@@ -1,34 +1,0 @@
-package com.x1.groo.forest.common.domain.aggregate;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Entity(name = "TempAnnouncementEntity")
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@ToString
-@Table(name = "announcement")
-public class AnnouncementEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Integer id;
-
-    @ManyToOne
-    @JoinColumn(name = "admin_id", nullable = false)
-    private UserEntity admin;
-
-    @Column(name = "title", nullable = false)
-    private String title;
-
-    @Column(name = "content", nullable = false, columnDefinition = "TEXT")
-    private String content;
-
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt = LocalDateTime.now();
-}
