--- src/main/java/com/x1/groo/forest/common/domain/aggregate/ItemEntity.java
@@ -1,32 +1,0 @@
-package com.x1.groo.forest.common.domain.aggregate;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity(name = "TempItemEntity")
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@ToString
-@Table(name = "item")
-public class ItemEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id", nullable = false)
-    private Integer id;
-
-    @Column(name = "name", nullable = false)
-    private String name;
-
-    @Column(name = "image_url", nullable = false)
-    private String imageUrl;
-
-    @ManyToOne
-    @JoinColumn(name = "category_id", nullable = false)
-    private CategoryEntity category;
-
-    @Column(name = "emotion", nullable = false)
-    private String emotion;
-}
