--- src/main/java/com/x1/groo/forest/common/domain/aggregate/PlacementEntity.java
@@ -1,35 +1,0 @@
-package com.x1.groo.forest.common.domain.aggregate;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.math.BigDecimal;
-
-@Entity(name = "TempPlacementEntity")
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@ToString
-@Table(name = "placement")
-public class PlacementEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="id")
-    private int id;
-
-    @Column(name = "position_x", precision = 9, scale = 6)
-    private BigDecimal positionX;
-
-    @Column(name = "position_y", precision = 9, scale = 6)
-    private BigDecimal positionY;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private UserEntity user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_item_id")
-    private UserItemEntity userItem;
-}
