+++ src/main/java/com/x1/groo/forest/common/domain/aggregate/UserEntity.java
@@ -5,7 +5,7 @@ import lombok.*;
 
 import java.time.LocalDateTime;
 
-@Entity(name = "TempUserEntity")
+@Entity(name = "ForestUserEntity")
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
