+++ src/main/java/com/x1/groo/forest/common/domain/repository/BackgroundRepository.java
@@ -4,6 +4,6 @@ import com.x1.groo.forest.common.domain.aggregate.BackgroundEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-@Repository("TempBackgroundRepository")
+@Repository
 public interface BackgroundRepository extends JpaRepository<BackgroundEntity, Integer> {
 }
