--- src/main/java/com/x1/groo/forest/common/domain/repository/MailboxRepository.java
@@ -1,15 +1,0 @@
-package com.x1.groo.forest.common.domain.repository;
-
-import com.x1.groo.forest.common.domain.aggregate.MailboxEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
-
-@Repository("TempMailboxRepository")
-public interface MailboxRepository extends JpaRepository<MailboxEntity, Integer> {
-    @Modifying
-    @Query("UPDATE MailboxEntity m SET m.isDeleted = true WHERE m.id = :mailboxId")
-    void softDeleteById(@Param("mailboxId") int mailboxId);
-}
