--- src/main/java/com/x1/groo/forest/common/domain/repository/PlacementRepository.java
@@ -1,12 +1,0 @@
-package com.x1.groo.forest.common.domain.repository;
-
-import com.x1.groo.forest.common.domain.aggregate.PlacementEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository("TempPlacementRepository")
-public interface PlacementRepository extends JpaRepository<PlacementEntity, Integer> {
-    void deleteByUserItemIdIn(List<Integer> userItemIds);
-}
