--- src/main/java/com/x1/groo/forest/emotion/command/domain/aggregate/ForestEntity.java
@@ -1,37 +1,0 @@
-package com.x1.groo.forest.emotion.command.domain.aggregate;
-
-import com.x1.groo.forest.mate.command.domain.aggregate.BackgroundEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@ToString
-@Table(name="forest")
-public class ForestEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="id")
-    private int id;
-
-    @Column(name="name")
-    private String name;
-
-    @Column(name="month")
-    private String month;
-
-    @Column(name="is_public")
-    private boolean isPublic;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="user_id")
-    private UserEntity user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="background_id")
-    private BackgroundEntity backgroundId;
-}
