--- src/main/java/com/x1/groo/forest/emotion/command/domain/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.x1.groo.forest.emotion.command.domain.repository;
-
-import com.x1.groo.forest.emotion.command.domain.aggregate.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository("ForestUserRepository")
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
-}
