--- src/main/java/com/x1/groo/forest/mate/command/domain/aggregate/BackgroundEntity.java
@@ -1,25 +1,0 @@
-package com.x1.groo.forest.mate.command.domain.aggregate;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@ToString
-@Table(name = "background")
-public class BackgroundEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private int id;
-
-    @Column(name = "name")
-    private String name;
-
-    @Column(name = "image_url")
-    private String imageUrl;
-}
